feat: back scrolly with real data + performance pass
Migrate the scrolly framework off synthetic nodes onto the real
~1k-actor sample (build-scrolly-nodes.js → src/data/scrolly-nodes.json),
with the per-edge tween model, networkIntro state, and node labels /
pulse ring.

Performance work validated against a throttled Pixel 7 trace (4x CPU),
which held under the 60fps budget on every frame at a 10k-node stress test:
- draw(): batch dots into one Path2D per (quantised rgb, alpha bucket)
  instead of a fillStyle + fill per dot
- tween: Float32Array for the per-frame hot arrays
- ScrollyVisual: memoize pure layouts by (state, w, h); position node
  labels via compositor-only transform/opacity rather than left/top

Also fix lint-staged cspell to pass --no-must-find-files so an
intentionally-ignored file (scrolly-nodes.json) no longer errors the hook.

+++ src/components/scrolly/tween.js
@@ -6,7 +6,7 @@ export const easeCubicInOut = (t) =>
 
 /**
  * @typedef {Object} Tweener
- * @property {Float64Array} current live rendered values
+ * @property {Float32Array} current live rendered values
  * @property {(next: Float64Array, ms: number, jitter?: number, nodeDelays?: Float64Array) => void} to
  * @property {() => void} stop
  */
@@ -24,8 +24,9 @@ export const easeCubicInOut = (t) =>
  */
 export function createTweener(size, draw, stride = 1) {
 	const groups = Math.ceil(size / stride);
-	const current = new Float64Array(size);
-	const start = new Float64Array(size);
+	// Float32 for the per-frame hot arrays; layout `target` stays Float64
+	const current = new Float32Array(size);
+	const start = new Float32Array(size);
 	const delays = new Float64Array(groups);
 	let target = null;
 	let startTime = 0;
